remove mulitple errors per test logic
