--- packages/runner/cypress/fixtures/errors/multiple_errors_spec.js
@@ -1,15 +1,0 @@
-import './setup'
-
-function fail() {
-  assert(false,`${this._runnable.hookName || this.test.title} - ${this._runnable.parent.title}`)
-}
-
-describe('multiple errors',() => {
-  describe('s1',() => {
-    afterEach(fail)
-    describe('s1.1',() => {
-      afterEach(fail)
-      it('t1',fail)
-    })
-  })
-})
