+++ packages/runner/cypress/integration/reporter.errors.spec.js
@@ -707,19 +707,4 @@ describe('errors ui', () => {
       method: 'cy.expect',
     })
   })
-
-  describe('multiple errors', {
-    // see more of the errors
-    viewportHeight: 1000,
-  }, () => {
-    const file = 'multiple_errors_spec.js'
-
-    verify.it('t1', {
-      file,
-      verifyFn () {
-        cy.get('.runnable-err').should('have.length', 3)
-        // TODO: write more assertions
-      },
-    })
-  })
 })
